Initial commit - working prototype
It works! The basic idea is that there's alawys section where one item
is being repeated. You find which number is in that slot at the moment
based on the current permutation index you're asking for, then you
remove that number, and run it through again without that item. You do
this until you get to 2 items and you can figure out which item goes
where based on if the index/offset you're asking for is odd or even.