Refactor algo to use scoped offsets
This works a bit differently than the previous version. It calculates
the index of the item that goes in the first slot for a theoretical set
of a certain length based on the offset. Then calculates the index for
a set that has one less element, and so on until it does it for a
2-item set. Once it has all the indexes, it loops through them and
extracts each item by its index from the original set, leaving us with
one element left over which it adds to the end.